fixed according pr commetns
