+++ angular.json
@@ -27,7 +27,8 @@
               "src/assets"
             ],
             "styles": [
-              "src/styles.scss"
+              "node_modules/reset-css/sass/_reset.scss",
+              "src/styles.scss"              
             ],
             "scripts": [],
             "es5BrowserSupport": true
