+++ src/app/app.component.scss
@@ -5,6 +5,7 @@
     flex-direction: column;
     justify-content: space-between;
 }
+
 .page-content {
     height: 100%;
     display: flex;
