+++ src/app/components/auth/auth.component.html
@@ -1,4 +1,4 @@
-<form [formGroup]="authForm" (submit)="logIn()" class="auth-form">
+<form [formGroup]="authForm" (submit)="logIn()" class="form">
   <p>Enter your login credentials</p>
   <input formControlName="login" class="form__input" type="text">
   <input formControlName="password" class="form__input" type="password">
