+++ src/app/components/auth/auth.component.scss
@@ -1,46 +1,54 @@
-@import "../../styles_lib/colors.scss";
-:host{
+@import "../../styles-lib/colors.scss";
+
+:host {
     display: flex;
     flex-direction: column;
     align-items: center;
 }
-.auth-form {
+
+.form {
     display: flex;
     flex-direction: column;
     align-items: center;
 
+    &__input {
+        margin: 10px 0;
+        width: 460px;
+        border: 2px solid $MainColor;
+        border-radius: 5px;
+        padding: 0 0 0 5px;
+        height: 30px;
+        font-size: 1.4rem;
+        color: $MainColor;
+        transition:.2s all ease;
+    }
+
+    &__submit {
+        height: 40px;
+        width: 100px;
+        border: 2px solid $SecondColor;
+        background-color: $MainColor;
+        color:#fff;
+        border-radius: 5px;
+        font-size: 1.4rem;
+    }
+
+    & p {
+        color: $MainColor;
+        font-size: 2rem;
+        font-weight: 600;
+    }
+
+    & input:focus {
+        outline: none;
+        box-shadow: 0 0 3px 3px $SecondColor;
+    }
+
 }
-.form__input {
-    margin:10px 0;
-    width: 460px;
-    border:2px solid $MainColor;
-    border-radius: 5px;
-    padding: 0 0 0 5px;
-    height: 30px;
-    font-size: 1.4rem;
-    color: $MainColor;
-    transition:.2s all ease;
-}
-.form__submit {
-    height: 40px;
-    width: 100px;
-    border:2px solid $SecondColor;
-    background-color: $MainColor;
-    color:#fff;
-    border-radius: 5px;
-    font-size: 1.4rem;
-}
+
 .error-message {
-    color:$SecondColor;
+    color: $SecondColor;
     font-size: 2rem;
     margin: 30px;
 }
-p {
-    color:$MainColor;
-    font-size: 2rem;
-    font-weight: 600;
-}
-input:focus {
-    outline:none;
-    box-shadow: 0 0 3px 3px $SecondColor;
-}
+
