+++ src/app/components/footer/footer.component.scss
@@ -1,4 +1,5 @@
-@import "../../styles_lib/colors.scss";
+@import "../../styles-lib/colors.scss";
+
 .footer {
     display: flex;
     align-items: center;
