+++ src/app/components/header/header.component.html
@@ -1,6 +1,6 @@
 <header class="header">
   <div class="header__logo">{{title}}</div>
   <div class="header__menu">
-    <div class="menu__item" *ngIf="authService.isAuth" (click)="logOut()">LogOut</div>
+    <div class="menu__item" *ngIf="isAuth" (click)="logOut()">LogOut</div>
   </div>
 </header>
