+++ src/app/components/header/header.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, Input } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { AuthService } from 'src/app/services/auth/auth.service';
 import { Router } from '@angular/router';
+import { Observable } from 'rxjs';
 
 
 
@@ -13,9 +14,15 @@ import { Router } from '@angular/router';
 export class HeaderComponent implements OnInit {
   @Input() title;
   
-  constructor(public authService:AuthService,private router:Router) { }
+  public isAuth:boolean;
+
+  constructor(private authService:AuthService,private router:Router) { }
 
   ngOnInit() {
+    this.authService.isAuth.subscribe(isAuth => {
+      this.isAuth = isAuth;      
+    });
+    this.authService.tryAuth();
   }
 
   logOut(){
