+++ src/app/models/auth-user.ts
@@ -1,0 +1,5 @@
+export class AuthUser {
+    login:string;
+    password:string;
+    token:string;
+}
