+++ src/app/models/stored-user.ts
@@ -1,0 +1,4 @@
+export class StoredUser {
+    login:string;
+    token:string;
+}
