+++ src/app/models/user.ts
@@ -1,0 +1,5 @@
+export class User {
+    login:string;
+    password:string;
+}
+
