--- src/app/models/user/user.ts
@@ -1,13 +1,0 @@
-export class User {
-    login:string;
-    password:string;
-}
-export class AuthUser {
-    login:string;
-    password:string;
-    token:string;
-}
-export class StoredUser {
-    login:string;
-    token:string;
-}
