+++ src/app/styles-lib/colors.scss
