+++ src/styles.scss
@@ -1,10 +1,11 @@
-* {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-}
 html {
     font-size: 14px;
     font-family: sans-serif;
     min-width: 480px;
+}
+
+body {
+    padding: 0;
+    margin: 0;
+    box-sizing: border-box;
 }
